refactor(shell): move E2E seam into desktop entry point
App.tsx no longer reads window.etherpadDesktop.e2eFlags. Shell now exports
a pure attachE2EHelpers(target) function and desktop's renderer/index.tsx
invokes it conditionally when the preload reports e2eFlags.enabled. Shell
src/ is now fully decoupled from runtime-specific globals.
