+++ packages/shell/src/index.ts
@@ -1,4 +1,5 @@
 export { App } from './App.js';
+export { attachE2EHelpers } from './e2e-seam.js';
 export {
   setPlatform,
   getPlatform,
